support reading of remote configuration data
*remote configuration data* are data about workspace deployments or
instances owned by the user making the request. The user is determined
by the key given in the request. Keys as used in this context are also
known as rerobots API tokens. The term *key* is motivated here because
there are plans to introduce other kinds of keys besides API tokens.

This changeset also includes small, sundry improvements.