implement connection mode: sshtun
Note that some of the robustness with respect to maintaining or
re-assigning tunnel hub associations is not provided here yet.
(Workspace deployments provided by rerobots already have it.)

The next significant change is to newly generate SSH server keys of
the Docker container at instantiation. (Until implemented, re-use of
keys in the image is a security bug.)