TEST: create pylint rcfile
Mostly has default values from

    pylint --generate-rcfile

In `make check`, pylint is called with `--exit-zero` because there are
many style warnings that can be addressed later and do not motivate
quitting before tests are executed.  Eventually, we may remove
`--exit-zero` if a good state of style is reached, together with a
good pylint rcfile.