CLI: default format is verbose, might not be YAML
Sometimes the output is YAML-compatible without the user explicitly
requesting `--format=YAML`.  To guarantee YAML output, now user must
explicitly request it.

This is motivated because sometimes it is better to make the output
formatted for terminal output being read by a human, which is an
objective not always aligned with YAML.