replace requests calls with aiohttp client
closes #8

In this changset, aiohttp client session is created during
instantiation of hardshare.api.HSAPIClient. It is closed automatically
during the class destructor (__del__).

The `run` method manages a client session internally that is distinct
from the class-level client session. It is not clear to me yet whether
these sessions should be the same to improve performance (because of
shared connection pools etc.). I do not expect it to be significant
because, in practice, advertising is just a matter of opening 1
WebSocket.