consistent use of term `token` (a.k.a. `API token`)
Before this changeset, the hardshare client referred to these mostly
as "keys". In practice, I find this to be confusing because there are
several distinct entities which are described as "keys", e.g., SSH
keys. My original motivation for calling API tokens "keys" was to
emphasize the importance of secrecy to users.