style: 1 blank line upper bound
This is the style used in the implementations of Rust
and Go, and it is the default style of rustfmt and gofmt.
This is different than Python (https://peps.python.org/pep-0008/#blank-lines),
which I overall prefer. However, for the sake of consistency,
because I am not certain that 2 blank lines is objectively
better, and because supporting this requires an unstable
option of rustfmt since > 5 years ago, in this changeset
we switch to 1 blank line.